extract Navigation API router to framework-neutral ES module
